requirements.txt
llama-index
llama-index-llms-ollama
SpeechRecognition
pyaudio
openai-whisper
pyttsx3